+++ graphannis/src/annis/db/relannis.rs
@@ -1147,7 +1147,8 @@ where
                                     updates.add_event(UpdateEvent::AddEdge {
                                         source_node: previous_token.clone(),
                                         target_node: created_token_id.to_string(),
-                                        component_type: "Ordering".to_string(),
+                                        component_type: AnnotationComponentType::Ordering
+                                            .to_string(),
                                         component_name: "relannis.text".to_string(),
                                         layer: ANNIS_NS.to_string(),
                                     })?;
@@ -1158,7 +1159,7 @@ where
                             updates.add_event(UpdateEvent::AddEdge {
                                 source_node: created_token_id.to_string(),
                                 target_node: next_token.to_string(),
-                                component_type: "Ordering".to_string(),
+                                component_type: AnnotationComponentType::Ordering.to_string(),
                                 component_name: "relannis.text".to_string(),
                                 layer: ANNIS_NS.to_string(),
                             })?;